Cleaned up and separated code into separate classes for gRPC services.
