+++ message-service/pom.xml
@@ -17,6 +17,12 @@
   </properties>
 
   <dependencies>
+    <dependency>
+      <groupId>com.sjchat</groupId>
+      <artifactId>general</artifactId>
+      <version>0.0.1-SNAPSHOT</version>
+      <scope>compile</scope>
+    </dependency>
 
     <dependency>
     <groupId>io.grpc</groupId>
